Move lazy mode functionality into a separate `useLazyQuery` hook
Enabling/disabling lazy mode is no longer handled by passing an
option into the `useQuery` hook. The `useLazyQuery` hook can
now be used when lazy mode is preferred. This gets us away from
using type inference to decide on the return type of `useQuery`.
